+++ src/main/java/net/fexcraft/mod/fvtm/sys/legacy/LandVehicle.java
@@ -674,7 +674,7 @@ public class LandVehicle extends GenericVehicle implements IEntityAdditionalSpaw
             double cir = ((WheelData)vehicle.getPart("left_front_wheel").getType().getInstallationHandlerData()).getRadius() * 2 * Static.PI;
             wheelsAngle += throttle * cir; if(wheelsAngle > 360) wheelsAngle -= 360; if(wheelsAngle < -360) wheelsAngle += 360;
         	vehicle.getAttribute("wheel_angle").setValue(wheelsAngle);
-        	vehicle.getAttribute("throttle").setValue(throttle);
+        	vehicle.getAttribute("throttle").setValue((float)throttle);
         }
         for(WheelEntity wheel : wheels){
             if(wheel != null && world != null){
